feat: 🎸 Added callback functions to setState and setProps
Added callback functions to setState and setProps function which receive
current widget.state and widget props as their first argument. This is
in addition to the old api, when an object is provided as a argument to
setState/setProps function, it works the same way as before. Only
function types are handled differently.