fix: 🐛 exclude assets without source (#115)
* fix: 🐛 exclude assets without source

Assets that don't have a source are pointless and should not be included
in server response. The problem with missing asset source can be
encountered when building multiple ES version with conditional build of
one (or more) ES version that require polyfill.js. Addressing this
problem in server integration pkg is better because it covers both SPA
and MPA integration.

* test: 💍 added test for server integration