refactor: 💡 Add support for JSON assets, improve assets loading
This commit introduces support for loading inline JSON and external JSON
files. The `loadJsonAssets` function was added to handle JSON-specific
logic, such as fetching, parsing and caching JSON data. Additionaly, the
asset loading logic for scripts and styles was refactored to improve
reusability and maintainability. The `loadAssets` function has been
extended to support loading JSON assets. Similar to scripts, styles can
now be marked as optional. If an optional style fails to load, a warning
is logged to the console, but no error is thrown. JSON assets are
treated as optional by default.

BREAKING CHANGE: 🧨 All load* functions can now return either a promise resolving to an
array of loading results or synchronously return the array of results
directly. If you are using async/await, this change likely does not
require any adjustments unless you rely on the exact order of results.
The order of results now strictly matches the order in which you
provided the array of assets to these functions. If any asset fails to
load for any reason, its corresponding value in the results array will
be null.

✅ Closes: usa-3681